vector-search/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.26.4
streamlit==1.35.0
